refactor(skills): skills_hub.rs 拆分为 skills_hub/ 目录模块
- 3062 行 → mod.rs ~150 行（Tauri command 壳）+ 11 个语义子模块
- core/fsutil/scan/registry/http/target_sync/lifecycle/repos/discover/updates_popular/usage
- 纯搬运；内部 fn/struct 字段升 pub(super)，测试经 mod.rs 重导出访问不变
- baseline 重生成 + 新文件登记豁免
验证：cargo test --lib skills_hub 16 passed；cargo check 零 error；新文件 rustfmt 过；gate 绿
